Wed 19/6 Fixing to Calibration
I've gotten the program to run further through the calibration process. I did this by fixing some references to headDatum, when it is actually called d, and by adding a setParameters function to the datum class.

Right now, its stuck on the calibration_finalise() function in the hydrosightModel.m because the drainage elevation created by the model_TFN_HMM is an empty vector.