Finding the issue on 29/7/24
I don't know how to fix this issue, the solve() function is not part of calibration, and thus skips over the definition for the objFn variable, however, without that definition, the function skips important parts. This is the case, even though we don't need the objFn variable in the solve function.