30/7 Changes
I had an epiphony while driving to uni this morning, and I realised why the error was occuring. It wasn't that the objFn was being skipped, its that objFn1 and objFn2 were being defined, but not objFn normally. So I changed it to work. Then I found another error with a use of Nan instead of NaN, and then lastly I found the next error which occurs during plotting, as ii+2 exceeds the limit which is 4. Why is it set at ii+2, and not just ii?