Fixes
Added a term called usepdf to HMM, which will set it to use either the pdf function or the Von Asmuth Function. I tried with the alternative version I calculated, and while it produced a somewhat viable answer, it didn't produce one I would be happy with. Noise.alpha was restricted to a max of 1, because values beyond a certain positive index were causing inf results. -1 was removed as the lower bound, because it wasn't the problem. I set ngs instead of popsize because it was defaulting to a population of 24 without the correct syntax.